Add files via upload
These files are meant to create an automated password recovery system through the use of security questions.  In my iteration I chose to add the ability to add the security questions through the profile page, and then created a link on the login page to start the process if the password has been forgotten.  The SecurityQuestion file is used solely to processes the data from ResetPage, and ConfirmReset serves the same role for the PasswordReset file.